+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- [(GH-167)](https://github.com/puppetlabs/pdkgo/issues/167) Implement `pct install` CLI command
+
 ## [0.3.0]
 
 - [(GH-144)](https://github.com/puppetlabs/pdkgo/issues/144) Implement `pct build` CLI command
-
-- [(GH-144)](https://github.com/puppetlabs/pdkgo/issues/144) Implement `pct build` CLI command
-- [(GH-167)](https://github.com/puppetlabs/pdkgo/issues/167) Implement `pct install` CLI command
 
 ### Removed